Remove yarn.lock
This is an automatically generated file with detailed version inforamation for all depencencies
in the OpenROV project.  Some of these dependencies have not been updated for a while and GitHub
is reporting possible security vulnerabilites.

By deleteing this file, the error messages should go away.  Note, this is only a temporary solution
as it does not address an actual update of the problematic dependencies.