Luminosity + Temperature Sensor Integration
This is the complete code base for running both the luminosity and temperature sensors, resolving earlier issues. In future I will work to streamline the code and remove relics/unneeded dependencies. As of this update both luminosity and temperature sensor print in Serial although the Cockpit was not working so I wasn't able to confirm with that.