Remove the Servo library from libraries/ directory
The Arduino Core includes a Servo library.  The one in the libraries/ directory
is redundant.  After reviwing both implementations, there is not much difference.
The implementaiton provided by the core is newer, but it does not provide any
difference in implementation. It support some additional architectures and uses
CPP standard types. It is an update of the same library.