Create a PinDefinitions.h file which uses the build macros
The PinDefinitions.h file is included by a number of the out of OpenRov supplied modules.
When splitting the file in to a surface and rover implementation, it is necessary to
split in a similar manner as SysConfig.h using  pre-processor macros.

This commit includes a PinDefinitions.h, which uses the pre-processor macros to decide
which module to include