Move some source files we want to keep for now
The first part of the code refactor required merging all of the code
for the surface and the rover.  There are a few files that have
conflicting code.  However, it is not entirely clear which version
of the file should be kept.

These files are stored in to the arduino_keep_for_now/ directory to allow
for further review.

NOTE: This directory should be considered temporary and will be deleted
after review.